backend/requirements.txt
aioredis==1.2.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
bcrypt==3.1.6
cffi==1.12.0
channels==2.1.7
channels-redis==2.3.3
constantly==15.1.0
cryptography==2.5
daphne==2.2.5
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.1
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
paramiko==2.4.2
psutil==5.5.1
pyasn1==0.4.5
pycparser==2.19
PyHamcrest==1.9.0
PyNaCl==1.3.0
pytz==2018.9
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
